Introduce Session + Page classes and rename Browser into Scraper
