+++ src/config.ts
@@ -1,0 +1,2 @@
+export const HOST = 'https://www.instagram.com';
+export const DEBUG = process.env.NODE_ENV !== 'production'
