+++ src/strategies/api.ts
@@ -1,0 +1,5 @@
+// TODO: xhr call to private rest api goes here
+
+export class RestApi {
+
+}
